+++ weldx/util.py
@@ -1,4 +1,6 @@
 """Contains package internal utility functions."""
+from __future__ import annotations
+
 import functools
 import json
 import sys
@@ -7,7 +9,7 @@ from collections.abc import Iterable, Sequence
 from functools import reduce, wraps
 from inspect import getmembers, isfunction
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Callable, Collection, Dict, List, Mapping, Union
+from typing import Any, Callable, Collection, Dict, List, Mapping, Union
 
 import numpy as np
 import pandas as pd
@@ -16,15 +18,11 @@ import xarray as xr
 from asdf.tags.core import NDArrayType
 from boltons import iterutils
 from pandas.api.types import is_datetime64_dtype, is_object_dtype, is_timedelta64_dtype
+from pint import DimensionalityError
 from scipy.spatial.transform import Rotation as Rot
 from scipy.spatial.transform import Slerp
 
-from weldx.constants import WELDX_QUANTITY as Q_
-from weldx.constants import WELDX_UNIT_REGISTRY as ureg
-from weldx.core import MathematicalExpression, TimeSeries
-
-if TYPE_CHECKING:  # pragma: no cover
-    import weldx.transformations as tf
+from .constants import WELDX_UNIT_REGISTRY as ureg
 
 
 class WeldxDeprecationWarning(DeprecationWarning):
@@ -228,74 +226,6 @@ def inherit_docstrings(cls):
     return cls
 
 
-def sine(
-    f: Union[pint.Quantity, str],
-    amp: Union[pint.Quantity, str],
-    bias: Union[pint.Quantity, str] = None,
-    phase: Union[pint.Quantity, str] = Q_(0, "rad"),
-) -> TimeSeries:
-    """Create a simple sine TimeSeries from quantity parameters.
-
-    f(t) = amp*sin(f*t+phase)+bias
-
-    Parameters
-    ----------
-    f :
-        Frequency of the sine (in Hz)
-    amp :
-        Sine amplitude
-    bias :
-        function bias
-    phase :
-        phase shift
-
-    Returns
-    -------
-    TimeSeries
-
-    """
-    if bias is None:
-        amp = Q_(amp)
-        bias = 0.0 * amp.u
-    expr_string = "a*sin(o*t+p)+b"
-    parameters = {"a": amp, "b": bias, "o": Q_(2 * np.pi, "rad") * Q_(f), "p": phase}
-    expr = MathematicalExpression(expression=expr_string, parameters=parameters)
-    return TimeSeries(expr)
-
-
-@deprecated(
-    "0.4.1",
-    "0.5.0",
-    "The 'LocalCoordinateSystem' now supports 'TimeSeries' as coordinates rendering "
-    "this function obsolete.",
-)
-def lcs_coords_from_ts(
-    ts: TimeSeries, time: Union[pd.DatetimeIndex, pint.Quantity]
-) -> xr.DataArray:
-    """Create translation coordinates from a TimeSeries at specific timesteps.
-
-    Parameters
-    ----------
-    ts:
-        TimeSeries that describes the coordinate motion as a 3D vector.
-    time
-        Timestamps used for interpolation.
-        TODO: add support for pd.DateTimeindex as well
-
-    Returns
-    -------
-    xarray.DataArray :
-        A DataArray with correctly labeled dimensions to be used for LCS creation.
-
-    """
-    ts_data = ts.interp_time(time=time).data_array
-    # assign vector coordinates and convert to mm
-    ts_data = ts_data.rename({"dim_1": "c"}).assign_coords({"c": ["x", "y", "z"]})
-    ts_data.data = ts_data.data.to("mm").magnitude
-    ts_data["time"] = pd.TimedeltaIndex(ts_data["time"].data)
-    return ts_data
-
-
 def is_column_in_matrix(column, matrix) -> bool:
     """Check if a column (1d array) can be found inside of a matrix.
 
@@ -385,7 +315,6 @@ def to_pandas_time_index(
         pd.TimedeltaIndex,
         pd.DatetimeIndex,
         xr.DataArray,
-        "tf.LocalCoordinateSystem",
     ],
 ) -> Union[pd.TimedeltaIndex, pd.DatetimeIndex]:
     """Convert a time variable to the corresponding pandas time index type.
@@ -407,16 +336,18 @@ def to_pandas_time_index(
 
     if isinstance(time, (pd.DatetimeIndex, pd.TimedeltaIndex)):
         return time
-
     if isinstance(time, LocalCoordinateSystem):
         return to_pandas_time_index(time.time)
 
     if isinstance(time, pint.Quantity):
+        time_ref = getattr(time, "time_ref", None)
         base = "s"  # using low base unit could cause rounding errors
         if not np.iterable(time):  # catch zero-dim arrays
             time = np.expand_dims(time, 0)
-        return pd.TimedeltaIndex(data=time.to(base).magnitude, unit=base)
-
+        delta = pd.TimedeltaIndex(data=time.to(base).magnitude, unit=base)
+        if time_ref is not None:
+            return delta + pd.Timestamp(time_ref)
+        return delta
     if isinstance(time, (xr.DataArray, xr.Dataset)):
         if "time" in time.coords:
             time = time.time
@@ -424,14 +355,15 @@ def to_pandas_time_index(
         if is_timedelta64_dtype(time_index) and time.weldx.time_ref:
             time_index = time_index + time.weldx.time_ref
         return time_index
-
-    if not np.iterable(time) or isinstance(time, str):
+    if (not np.iterable(time) or isinstance(time, str)) and not isinstance(
+        time, np.ndarray
+    ):
         time = [time]
+
     time = pd.Index(time)
 
     if isinstance(time, (pd.DatetimeIndex, pd.TimedeltaIndex)):
         return time
-
     # try manual casting for object dtypes (i.e. strings), should avoid integers
     # warning: this allows something like ["1","2","3"] which will be ns !!
     if is_object_dtype(time):
@@ -444,32 +376,6 @@ def to_pandas_time_index(
     raise TypeError(
         f"Could not convert {_input_type} " f"to pd.DatetimeIndex or pd.TimedeltaIndex"
     )
-
-
-def pandas_time_delta_to_quantity(
-    time: pd.TimedeltaIndex, unit: str = "s"
-) -> pint.Quantity:
-    """Convert a `pandas.TimedeltaIndex` into a corresponding `pint.Quantity`.
-
-    Parameters
-    ----------
-    time : pandas.TimedeltaIndex
-        Instance of `pandas.TimedeltaIndex`
-    unit :
-        String that specifies the desired time unit.
-
-    Returns
-    -------
-    pint.Quantity :
-        Converted time quantity
-
-    """
-    # from pandas Timedelta documentation: "The .value attribute is always in ns."
-    # https://pandas.pydata.org/pandas-docs/version/0.23.4/generated/pandas.Timedelta.html
-    nanoseconds = time.values.astype(np.int64)
-    if len(nanoseconds) == 1:
-        nanoseconds = nanoseconds[0]
-    return Q_(nanoseconds, "ns").to(unit)
 
 
 def matrix_is_close(mat_a, mat_b, abs_tol=1e-9) -> bool:
@@ -918,6 +824,12 @@ def xr_check_coords(dax: xr.DataArray, ref: dict) -> bool:
     ``optional`` : boolean
         default ``False`` - if ``True``, the dimension has to be in the DataArray dax
 
+    ``dimensionality`` : str or pint.Unit
+        Check if ``.attrs["units"]`` is the requested dimensionality
+
+    ``units`` : str or pint.Unit
+        Check if ``.attrs["units"]`` matches the requested unit
+
     Parameters
     ----------
     dax : xarray.DataArray
@@ -994,6 +906,27 @@ def xr_check_coords(dax: xr.DataArray, ref: dict) -> bool:
             ):
                 raise TypeError(
                     f"Mismatch in the dtype of the DataArray and ref['{key}']"
+                )
+
+        if "units" in check:
+            units = coords[key].attrs.get("units", None)
+            if not units or not (ureg.Unit(units) == ureg.Unit(check["units"])):
+                raise ValueError(
+                    f"Unit mismatch in coordinate '{key}'\n"
+                    f"Coordinate has unit '{str(units)}', expected '{check['units']}'"
+                )
+
+        if "dimensionality" in check:
+            units = coords[key].attrs.get("units", None)
+            dim = check["dimensionality"]
+            if not units or not (
+                ureg.get_dimensionality(units) == ureg.get_dimensionality(dim)
+            ):
+                raise DimensionalityError(
+                    units,
+                    check["dimensionality"],
+                    f"\nDimensionalit mismatch in coordinate '{key}'\n"
+                    f"Coordinate has unit '{str(units)}', expected '{dim}'",
                 )
 
     return True