+++ CHANGELOG.md
@@ -4,14 +4,21 @@
 
 ### added
 
+- added "units" (exact) and "dimensionality" (dimensionality compatible) checking options
+  to `util.xr_check_coords` [[#442]](https://github.com/BAMWelDX/weldx/pull/442)
+- `Time` class that can be initialized from several other time types and provides time related utility functions 
+  [[#433]](https://github.com/BAMWelDX/weldx/pull/433)
+
 ### removed
 
 ### changes
 
+- move `sine` utility function to `weldx.welding.util` [[#439]](https://github.com/BAMWelDX/weldx/pull/439)
+
 ### fixes
 
-- `WeldxFile.show_asdf_header` prints output on console, before it only returned the header
-  as parsed dict and string representation [[#428]](https://github.com/BAMWelDX/weldx/pull/428).
+- `WeldxFile.show_asdf_header` prints output on console, before it only returned the header as parsed dict and string
+  representation [[#428]](https://github.com/BAMWelDX/weldx/pull/428).
 
 ### documentation